requirements-dev.txt
django-debug-toolbar==3.1.1
django-extensions==3.0.9
Werkzeug==1.0.1
pre-commit==2.7.1
